+++ Makefile.mingw
@@ -33,10 +33,10 @@ nanodump: clean
 	$(CC_x64) -c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.o $(OPTIONS) $(SSP_OPTIONS) -DBOF -DSSP
 	$(STRIP_x64) --strip-unneeded dist/$(BOFNAME)_ssp.x64.o
 
-	$(CC_x64) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP
+	$(CC_x64) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/ssp_utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP
 	$(STRIP_x64) --strip-all dist/$(BOFNAME)_ssp.x64.exe
 
-	$(CC_x86) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x86.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP
+	$(CC_x86) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/ssp_utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x86.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP
 	$(STRIP_x86) --strip-all dist/$(BOFNAME)_ssp.x86.exe
 
 	$(CC_x64) -c source/delete_file.c -o dist/delete_file.x64.o $(OPTIONS) -DBOF
@@ -91,9 +91,9 @@ debug: clean
 
 	$(CC_x64) -c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.o $(OPTIONS) $(SSP_OPTIONS) -DBOF -DSSP -DDEBUG
 
-	$(CC_x64) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP -DDEBUG
+	$(CC_x64) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/ssp_utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x64.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP -DDEBUG
 
-	$(CC_x86) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x86.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP -DDEBUG
+	$(CC_x86) source/utils.c source/syscalls.c source/dinvoke.c source/pipe.c source/ssp/ssp_utils.c source/ssp/ssp.c -o dist/$(BOFNAME)_ssp.x86.exe $(OPTIONS) $(SSP_OPTIONS) -DEXE -DSSP -DDEBUG
 
 	$(CC_x64) -c source/delete_file.c -o dist/delete_file.x64.o $(OPTIONS) -DBOF -DDEBUG
